Add shebang bash with the env utility
This commit is to introduce to the shebang (e.g., #!) bash with the 'env'
utility to tell the operating system whitch interpreter to use to parse
the rest of the file.

There are two ways to use the Shebang directive and set the interpreter.
01. Using the absolute path to the bash binary
 #!/bin/bash

02. Using the env utility
 #!/usr/bin/env bash

The advantage of using the second approach is that it will search for the bash
executable in the user’s $PATH environmental variable. If there are more than
one paths to bash, the first one will be used by the script.

(I AGREE this CLA.)

Signed-off-by: Geunsik Lim <leemgs@gmail.com>