Add automated code review list
This commit is to append the CODEOWNERS file to define code review list
by default that is responsible for code in this repository.
After this PR, code owners are automatically requested for code review
when a someone submits a pull request (PR) that add or modifies a source code
that they own.

Signed-off-by: Geunsik Lim <leemgs@gmail.com>